requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2025.4
cssselect2==0.8.0
Django==5.2
django-crispy-forms==2.4
django-extensions==4.1
django-jalali==7.4.0
django-jalali-date==2.0.0
Django-jalali-date-new==1.1
django-prometheus==2.3.1
django-pwa==2.0.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
fonttools==4.58.0
google==3.0.0
gunicorn==23.0.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
kombu==5.5.3
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
prometheus_client==0.22.0
prompt_toolkit==3.0.51
protobuf==3.20.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2025.4
cssselect2==0.8.0
Django==5.2
django-crispy-forms==2.4
django-extensions==4.1
django-jalali==7.4.0
django-jalali-date==2.0.0
Django-jalali-date-new==1.1
django-prometheus==2.3.1
django-pwa==2.0.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
fonttools==4.58.0
google==3.0.0
gunicorn==23.0.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
kombu==5.5.3
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
prometheus_client==0.22.0
prompt_toolkit==3.0.51
protobuf==3.20.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2025.4
cssselect2==0.8.0
Django==5.2
django-crispy-forms==2.4
django-extensions==4.1
django-jalali==7.4.0
django-jalali-date==2.0.0
Django-jalali-date-new==1.1
django-prometheus==2.3.1
django-pwa==2.0.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
fonttools==4.58.0
google==3.0.0
gunicorn==23.0.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
kombu==5.5.3
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
prometheus_client==0.22.0
prompt_toolkit==3.0.51
protobuf==3.20.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
